pyproject.toml
[tool.poetry]
name = "product-need-request-system"
version = "0.1.0"
description = "System for managing product need requests, proposals, lifecycles, supplier interaction, and automated evaluations."
authors = ["AI Agent <ai@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = {version = "^2.0.23", extras = ["postgresql_psycopg2binary"]}
pydantic = {extras = ["email"], version = "^2.5.2"}
alembic = "^1.13.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6" # For form data, e.g. OAuth2 password flow
tenacity = "^8.2.3" # For retrying operations, e.g. DB connection
# Dependencies for Celery + Redis (will be used in a later step)
# celery = {extras = ["redis"], version = "^5.3.6"}
# redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
httpx = "^0.25.1" # For testing FastAPI endpoints

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
